解决一些高危漏洞

解决了两类问题：
1）异常状态下的内存泄露
2）可能越界的风险

See merge request !1382238

+++ src/congestion_control/xqc_cubic_kernel.c
@@ -560,7 +560,7 @@ static uint64_t
 xqc_cubic_get_cwnd (void *cong_ctl)
 {
     xqc_cubic_kernel_t *ca = (xqc_cubic_kernel_t*)(cong_ctl);
-    return ca->cwnd * XQC_CUBIC_MSS;
+    return (uint64_t)ca->cwnd * XQC_CUBIC_MSS;
 }
